requirements.txt
# Core RAG Framework
llama-index-core
llama-index
llama-index-retrievers-bm25

# Vector Database
llama-index-vector-stores-qdrant
qdrant-client

# Graph Database
llama-index-graph-stores-neo4j
neo4j

# LLM Providers
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-llms-ollama
llama-index-embeddings-ollama

# Multi-Agent Orchestration
langgraph
langchain
langchain-core

# Utilities
python-dotenv

# Data Sources
requests
